ref: Remove aorta completely
