+++ Cargo.toml
@@ -32,9 +32,6 @@ serde_derive = "1.0.72"
 serde = "1.0.72"
 serde_json = "1.0.26"
 
-[dependencies.semaphore-aorta]
-path = "aorta"
-
 [dependencies.semaphore-common]
 path = "common"
