--- aorta/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-authors = ["Sentry <hello@sentry.io>"]
-description = "The relay <-> backend communcation and authentication protocol"
-documentation = "https://docs.rs/semaphore-common"
-homepage = "https://github.com/getsentry/semaphore"
-license = "MIT"
-name = "semaphore-aorta"
-repository = "https://github.com/getsentry/semaphore"
-version = "0.1.3"
-
-[dependencies]
-base64 = "0.9.2"
-failure = "0.1.2"
-failure_derive = "0.1.2"
-futures = "0.1.23"
-hyper = "0.11.27"
-hyper-tls = "0.1.3"
-parking_lot = "0.6.3"
-rand = "0.5.5"
-serde = "1.0.72"
-serde_derive = "1.0.72"
-serde_json = "1.0.26"
-tokio-core = "0.1.17"
-url = "1.7.1"
-log = "0.4.4"
-rust_sodium = "0.9.0"
-url_serde = "0.2.0"
-base64-serde = "0.3.0"
-
-[dependencies.chrono]
-features = ["serde"]
-version = "0.4.5"
-
-[dependencies.semaphore-common]
-path = "../common"
-
-[dependencies.uuid]
-features = ["v4", "serde"]
-version = "0.6.5"
