--- aorta/src/lib.rs
@@ -1,43 +1,0 @@
-//! Implements the relay <-> backend protocol.
-#![warn(missing_docs)]
-extern crate base64;
-#[macro_use]
-extern crate base64_serde;
-extern crate chrono;
-extern crate failure;
-#[macro_use]
-extern crate failure_derive;
-extern crate futures;
-extern crate hyper;
-#[macro_use]
-extern crate log;
-extern crate parking_lot;
-extern crate rand;
-extern crate rust_sodium;
-#[macro_use]
-extern crate semaphore_common;
-extern crate serde;
-#[macro_use]
-extern crate serde_derive;
-extern crate serde_json;
-extern crate tokio_core;
-extern crate url;
-extern crate url_serde;
-extern crate uuid;
-
-mod api;
-mod auth;
-mod config;
-mod event;
-mod projectstate;
-mod query;
-mod upstream;
-mod utils;
-
-pub use api::*;
-pub use auth::*;
-pub use config::*;
-pub use event::*;
-pub use projectstate::*;
-pub use query::*;
-pub use upstream::*;
