+++ common/Cargo.toml
@@ -9,17 +9,36 @@ repository = "https://github.com/getsentry/semaphore"
 version = "0.1.3"
 
 [dependencies]
+base64 = "0.9.2"
+# base64-serde = "0.3.0"
+cadence = "0.15.1"
+chrono = "0.4.5"
 failure = "0.1.2"
 failure_derive = "0.1.2"
-url = "1.7.1"
-serde = "1.0.70"
-serde_derive = "1.0.70"
-serde_json = "1.0.24"
-hyper = "0.11.27"
-sentry-types = "0.6.2"
 lazy_static = "1.0.2"
-cadence = "0.15.1"
-parking_lot = "0.6.3"
-log = "0.4.3"
-marshal = { git = "https://github.com/getsentry/marshal", rev = "810d521dfbde1e2d0c1e63fbb2e226ef149a8883" }
+log = "0.4.4"
 human-size = "0.4.0"
+marshal = { git = "https://github.com/getsentry/marshal", rev = "810d521dfbde1e2d0c1e63fbb2e226ef149a8883" }
+parking_lot = "0.6.3"
+rand = "0.5.5"
+rust_sodium = "0.9.0"
+sentry-types = "0.6.2"
+serde = "1.0.74"
+serde_derive = "1.0.74"
+serde_json = "1.0.26"
+url = "1.7.1"
+uuid = { version = "0.6.5", features = ["v4", "serde"] }
+
+# failure = "0.1.2"
+# failure_derive = "0.1.2"
+# futures = "0.1.23"
+# hyper = "0.11.27"
+# hyper-tls = "0.1.3"
+# log = "0.4.4"
+# parking_lot = "0.6.3"
+# serde = "1.0.72"
+# serde_derive = "1.0.72"
+# serde_json = "1.0.26"
+# tokio-core = "0.1.17"
+# url = "1.7.1"
+# url_serde = "0.2.0"
