+++ config/Cargo.toml
@@ -12,20 +12,14 @@ version = "0.1.3"
 chrono = "0.4.5"
 failure = "0.1.2"
 failure_derive = "0.1.2"
+log = { version = "0.4.4", features = ["serde"] }
 sentry = "0.8.0"
-serde = "1.0.72"
-serde_derive = "1.0.72"
+serde = "1.0.74"
+serde_derive = "1.0.74"
 serde_json = "1.0.26"
-serde_yaml = "0.8.0"
+serde_yaml = "0.8.1"
 url = "1.7.1"
 url_serde = "0.2.0"
-
-[dependencies.log]
-features = ["serde"]
-version = "0.4.3"
-
-[dependencies.semaphore-aorta]
-path = "../aorta"
 
 [dependencies.semaphore-common]
 path = "../common"
