+++ server/Cargo.toml
@@ -32,14 +32,11 @@ openssl = { version = "0.10.11", optional = true }
 regex = "1.0.2"
 sentry = "0.8.0"
 sentry-actix = "0.3.0"
-serde = "1.0.72"
-serde_derive = "1.0.72"
+serde = "1.0.74"
+serde_derive = "1.0.74"
 serde_json = "1.0.26"
 url = "1.7.1"
 uuid = { version = "0.6.5", features = ["v4"] }
-
-[dependencies.semaphore-aorta]
-path = "../aorta"
 
 [dependencies.semaphore-config]
 path = "../config"
@@ -48,4 +45,4 @@ path = "../config"
 path = "../common"
 
 [target."cfg(not(windows))".dependencies]
-libc = "0.2.42"
+libc = "0.2.43"
