+++ server/src/actors/upstream.rs
@@ -11,10 +11,8 @@ use futures::prelude::*;
 use serde::de::DeserializeOwned;
 use serde::ser::Serialize;
 
-use semaphore_aorta::{
-    RegisterChallenge, RegisterRequest, RegisterResponse, Registration, UpstreamDescriptor,
-};
-use semaphore_config::Credentials;
+use semaphore_common::{RegisterChallenge, RegisterRequest, RegisterResponse, Registration};
+use semaphore_config::{Credentials, UpstreamDescriptor};
 
 #[derive(Fail, Debug)]
 pub enum UpstreamRequestError {
