+++ server/src/utils.rs
@@ -1,5 +1,8 @@
+use std::fmt;
+
 use actix::dev::{MessageResponse, ResponseChannel};
 use actix::prelude::*;
+use failure::Fail;
 use futures::prelude::*;
 
 pub enum Response<T, E> {
@@ -76,6 +79,55 @@ where
                     tx.send(res);
                 }
             }
+        }
+    }
+}
+
+/// An error response from an api.
+#[derive(Serialize, Deserialize, Default, Debug)]
+pub struct ApiErrorResponse {
+    detail: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    causes: Option<Vec<String>>,
+}
+
+impl ApiErrorResponse {
+    /// Creates an error response with a detail message
+    pub fn with_detail<S: AsRef<str>>(s: S) -> ApiErrorResponse {
+        ApiErrorResponse {
+            detail: Some(s.as_ref().to_string()),
+            causes: None,
+        }
+    }
+
+    /// Creates an error response from a fail.
+    pub fn from_fail<F: Fail>(fail: &F) -> ApiErrorResponse {
+        let mut messages = vec![];
+
+        for cause in Fail::iter_chain(fail) {
+            let msg = cause.to_string();
+            if !messages.contains(&msg) {
+                messages.push(msg);
+            }
+        }
+
+        ApiErrorResponse {
+            detail: Some(messages.remove(0)),
+            causes: if messages.is_empty() {
+                None
+            } else {
+                Some(messages)
+            },
+        }
+    }
+}
+
+impl fmt::Display for ApiErrorResponse {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        if let Some(ref detail) = self.detail {
+            write!(f, "{}", detail)
+        } else {
+            write!(f, "no error details")
         }
     }
 }
